Remove custom block executor
