+++ crates/subcoin-node/src/cli.rs
@@ -84,7 +84,6 @@ pub fn run() -> sc_cli::Result<()> {
             })
         }
         Command::ImportBlocks(cmd) => {
-            let block_execution_strategy = cmd.common_params.block_execution_strategy();
             let bitcoin_network = cmd.common_params.bitcoin_network();
             let import_config = ImportConfig {
                 execute_block: cmd.execute_transactions,
@@ -101,12 +100,10 @@ pub fn run() -> sc_cli::Result<()> {
                 let subcoin_service::NodeComponents {
                     client,
                     task_manager,
-                    block_executor,
                     ..
                 } = subcoin_service::new_node(subcoin_service::SubcoinConfiguration {
                     network: bitcoin_network,
                     config: &config,
-                    block_execution_strategy,
                     no_hardware_benchmarks: true,
                     storage_monitor,
                 })?;
@@ -128,7 +125,6 @@ pub fn run() -> sc_cli::Result<()> {
                 Ok((
                     import_blocks_cmd.run(
                         client,
-                        block_executor,
                         data_dir,
                         import_config,
                         spawn_handle,
@@ -140,7 +136,6 @@ pub fn run() -> sc_cli::Result<()> {
         }
         Command::Tools(tools) => tools.run(),
         Command::Blockchain(blockchain) => {
-            let block_execution_strategy = blockchain.block_execution_strategy();
             let cmd = BlockchainCmd::new(blockchain);
             let runner = SubstrateCli.create_runner(&cmd)?;
             runner.async_run(|config| {
@@ -151,7 +146,6 @@ pub fn run() -> sc_cli::Result<()> {
                 } = subcoin_service::new_node(subcoin_service::SubcoinConfiguration {
                     network: bitcoin::Network::Bitcoin,
                     config: &config,
-                    block_execution_strategy,
                     no_hardware_benchmarks: true,
                     storage_monitor,
                 })?;
